Create Interface for all scraper related fucntions

+++ StockScrapApi/Models/BasicInfo.cs
@@ -10,11 +10,11 @@ namespace StockScrapApi.Models
     public class BasicInfo
     {
         public int Id { get; set; }
-        public float? AuthorizedCapital { get; set; }
+        public double? AuthorizedCapital { get; set; }
         public string DebutTradingDate { get; set; }
-        public float? PaidUpCapital { get; set; }
-        public float? FaceParValue { get; set; }
-        public float? TotalOutstandingSecurity { get; set; }
+        public double? PaidUpCapital { get; set; }
+        public double? FaceParValue { get; set; }
+        public double? TotalOutstandingSecurity { get; set; }
         public string InstrumentType { get; set; }
         public int? MarketLot { get; set; }
         public string Sector { get; set; }
